Extended key codes for arrow keys

+++ src-spec-next/ps2comm.h
@@ -6,8 +6,8 @@ void ps2c_init(uint pin_clk, uint pin_dat);
 // call this once per millisecond
 int ps2c_tick(void); 
 
-int ps2c_keydown(uint8_t code);
-int ps2c_keyup(uint8_t code);
+int ps2c_keydown(uint16_t code);
+int ps2c_keyup(uint16_t code);
 
 uint8_t ps2c_getleds(void);
